Created COntactList Rough Final
This version fulfills the requirements for all the cases.  Im going to work on the 1a case tomorrow, if anyone else feels like taking a shot go ahead, all the 1a requirements are just input checks.  Please run the full program and try to find any bugs that might still be present.